+++ testdata/page.html
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Shitposting 9001</title>
+  <meta name="description" content="Shitposting Images" />
+  <!--<meta name="keywords" content="thumbnails, grid, preview, google image search, jquery, image grid, expanding, preview, portfolio" />-->
+  <meta name="author" content="sp9k1" />
+  <link rel="shortcut icon" href="../favicon.ico">
+  <link rel="stylesheet" type="text/css" href="static/default.css" />
+  <link rel="stylesheet" type="text/css" href="static/component.css" />
+  <script src="static/modernizr.custom.js"></script>
+</head>
+
+<body>
+  <div class="container">
+    <header class="clearfix">
+      <h1>Shitposting 9001 <span>with more shitposting</span></h1>
+    </header>
+    <div class="main">
+      <ul id="og-grid" class="og-grid">
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/6_gummy_bears.png" data-title="6 gummy bears" data-description="Archer drugs">
+            <img width="250px" height="250px" src="/6_gummy_bears.png" alt="6 gummy bears" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/dickbutt_for_mayor.gif" data-title="dickbutt for mayor" data-description="dickbutt">
+            <img width="250px" height="250px" src="/dickbutt_for_mayor.gif" alt="dickbutt for mayor" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/fort_kickass.jpg" data-title="fort kickass" data-description="Archer work">
+            <img width="250px" height="250px" src="/fort_kickass.jpg" alt="fort kickass" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/fuck_you_space.png" data-title="fuck you space" data-description="Archer space">
+            <img width="250px" height="250px" src="/fuck_you_space.png" alt="fuck you space" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/janis_joplin_rimjob.jpg" data-title="janis joplin rimjob" data-description="Archer">
+            <img width="250px" height="250px" src="/janis_joplin_rimjob.jpg" alt="janis joplin rimjob" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/lemur_pudding_cups.jpg" data-title="lemur pudding cups" data-description="Archer woodhouse">
+            <img width="250px" height="250px" src="/lemur_pudding_cups.jpg" alt="lemur pudding cups" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/looking_for_a_terrorist_and_ocelot.gif" data-title="looking for a terrorist and ocelot" data-description="Archer terrorist">
+            <img width="250px" height="250px" src="/looking_for_a_terrorist_and_ocelot.gif" alt="looking for a terrorist and ocelot" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/ninjas_are_sexy.gif" data-title="ninjas are sexy" data-description="Archer ninjas">
+            <img width="250px" height="250px" src="/ninjas_are_sexy.gif" alt="ninjas are sexy" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/sour_mix_in_margarits.png" data-title="sour mix in margaritas" data-description="Archer alcohol">
+            <img width="250px" height="250px" src="/sour_mix_in_margarits.png" alt="sour mix in margaritas" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/spooning_a_barret.png" data-title="spooning a barret" data-description="Archer gun">
+            <img width="250px" height="250px" src="/spooning_a_barret.png" alt="spooning a barret" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/throbbing_erection.png" data-title="throbbing erection" data-description="Archer">
+            <img width="250px" height="250px" src="/throbbing_erection.png" alt="throbbing erection" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/whats_in_the_case.gif" data-title="whats in the case" data-description="Archer train">
+            <img width="250px" height="250px" src="/whats_in_the_case.gif" alt="whats in the case" />
+          </a>
+        </li>
+        <li>
+          <a href="https://github.com/jakdept/sp9k1/" data-largesrc="/woodhouse_like_people.png" data-title="woodhouse like people" data-description="Archer woodhouse">
+            <img width="250px" height="250px" src="/woodhouse_like_people.png" alt="woodhouse like people" />
+          </a>
+        </li>
+      </ul>
+    </div>
+  </div>
+  <!-- /container -->
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+  <script src="static/grid.js"></script>
+  <script>
+    $(function () {
+      Grid.init();
+    });
+  </script>
+</body>
+
+</html>
